swapped QtyAccmps and Announcements between ProjectTender and ProjectPlan
